requirements.txt
alembic
APScheduler==3.10.2
Babel==2.12.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
